// menus-front/src/app/shared/models/week.model.ts
import { WeekMeal } from './week-meal.model';

export class Week {
    id: string;
    mondayLunch: WeekMeal;
    mondayDinner: WeekMeal;
    tuesdayLunch: WeekMeal;
    tuesdayDinner: WeekMeal;
    wednesdayLunch: WeekMeal;
    wednesdayDinner: WeekMeal;
    thursdayLunch: WeekMeal;
    thursdayDinner: WeekMeal;
    fridayLunch: WeekMeal;
    fridayDinner: WeekMeal;
    saturdayLunch: WeekMeal;
    saturdayDinner: WeekMeal;
    sundayLunch: WeekMeal;
    sundayDinner: WeekMeal;
}